seprate request and response struct from service task
